API-12088 remove common objects from webhooks and adjust structures
