+++ src/webhooks/structures/index.ts
@@ -1,0 +1,2 @@
+export * from "./events";
+export * from "./structures";
